+++ libyul/YulStack.cpp
@@ -31,6 +31,7 @@
 #include <libyul/optimiser/Semantics.h>
 #include <libyul/optimiser/Suite.h>
 #include <libevmasm/Assembly.h>
+#include <libevmasm/CodeSizeLimits.h>
 #include <libevmasm/Ethdebug.h>
 #include <liblangutil/Scanner.h>
 #include <liblangutil/SourceReferenceFormatter.h>
@@ -296,6 +297,15 @@ YulStack::assembleWithDeployed(std::optional<std::string_view> _deployName, bool
 		reportUnimplementedFeatureError(_error);
 		return {MachineAssemblyObject{}, MachineAssemblyObject{}};
 	}
+
+	yulAssert(creationObject.bytecode);
+	evmasm::checkCodeSizeLimits(
+		m_errorReporter,
+		m_evmVersion,
+		m_parserResult->code()->root().debugData->nativeLocation,
+		creationObject.bytecode->bytecode.size(),
+		deployedObject.bytecode ? deployedObject.bytecode->bytecode.size() : 0
+	);
 
 	return {std::move(creationObject), std::move(deployedObject)};
 }